--- hangman.js
@@ -1,18 +1,0 @@
-const readline = require('readline-sync');
-
-function getValidLetterGuess() {
-  function guessIsValid(letter) {
-    return letter.length === 1 && letter.toUpperCase() != letter.toLowerCase()
-  }
-  let letter = ""
-  while (!letter) {
-    let input = readline.question("Please enter your guess: ")
-    if (guessIsValid(input)) {
-      letter = input
-    } else {
-      console.log("Please enter a valid letter")
-    }
-  }
-  return letter.toLowerCase()
-}
-
